Move cronjobs to separate main method
