+++ src/orukomm/logic/scheduler/jobs/Email.java
@@ -1,4 +1,4 @@
-package orukomm.logic;
+package orukomm.logic.scheduler.jobs;
 
 import java.util.ArrayList;
 import java.util.Properties;
@@ -82,7 +82,7 @@ public class Email {
             message.setContent(body, "text/html; charset=utf8");
             
             Transport.send(message);
-            System.out.println("E-mail successfully sent");
+            System.out.println("E-mails successfully sent");
 
         } catch (MessagingException me) {
             throw new RuntimeException(me);
